Semana_12/flask_api/requirements.txt
Flask
PyJWT
Flask-Cors
